move upload field
