+++ src/packages/property-editors/upload-field/manifests.ts
