+++ src/packages/property-editors/upload-field/property-editor-ui-upload-field.stories.ts
